feat: Explicit autowiring of crates
On a multi-crate workspace, disable autowiring of crates by default. Let the user wire the individual crates manually. Otherwise, they will be forced to configure each crate to build in Nix separately (via specifying `buildInputs`) which is quite unnecessary if only a handful of crates need to be exposed in the flake.